Step 8: Cleanup
- Remove `counter_backend.rs` and `virtual_backend.rs`
- Remove `event_library.rs` and `event_registry.rs` (merged into
`event.rs`)
- Update `lib.rs` module declarations
- Update imports
