+++ src/main.rs
@@ -1,8 +1,8 @@
 use clap::Parser;
 use indicatif::{ProgressBar, ProgressStyle};
 use saccade::cli::{Cli, Commands};
-use saccade::event_library::EventLibrary;
-use saccade::event_registry::EventRegistry;
+use saccade::event::EventLibrary;
+use saccade::event::EventRegistry;
 use saccade::perf::Perf;
 use saccade::perfetto::{self, PerfettoWriter};
 use saccade::profiler::ProfilerBuilder;
@@ -16,7 +16,7 @@ use saccade::source::SampleSource;
 use saccade::source::hardware::HardwareSampleSource;
 use saccade::source::virtual_source::VirtualSampleSource;
 use saccade::syscalls;
-use saccade::virtual_backend::TimeVaryingRates;
+use saccade::source::virtual_source::TimeVaryingRates;
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::BufReader;
