+++ src/scheduler/round_robin.rs
@@ -1,4 +1,4 @@
-use crate::event_registry::EventId;
+use crate::event::EventId;
 use crate::quantum::Quantum;
 use crate::scheduler::ScheduleDecision;
 use crate::scheduler::Scheduler;
